pdf-rag-system/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.1
python-multipart>=0.0.12
pydantic>=2.11.7
python-dotenv>=1.1.0
sqlalchemy>=2.0.35
pymysql>=1.1.0
alembic>=1.13.2
cryptography>=46.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
akshare>=1.18.30
python-dateutil>=2.9.0
tushare>=1.4.0
redis>=5.0.0
openai>=1.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
pymupdf4llm>=0.0.17
pdfplumber>=0.11.0
Pillow>=10.0.0
pytesseract>=0.3.10
